looks like auth is working now

+++ GdanskExplorer/Data/Trip.cs
@@ -7,7 +7,7 @@ public class Trip
     public Guid Id { get; set; }
     public User User { get; set; } = null!;
     public MultiPolygon Polygon { get; set; } = null!;
-    public float Area { get; set; }
-    public float Length { get; set; }
+    public double Area { get; set; }
+    public double Length { get; set; }
     public DateTime UploadDate { get; set; }
 }
